Фикс к HandExtinguish
Обнимания происходят и на стороне клиента - поэтому система перенесена в shared
